# Wait for ueventd to create loop device on Android