Reference: https://cs.android.com/android/_/android/platform/external/toybox/+/51a43ad5225153582b40d3fd289701efc63c8f62