@tiann 2023-02-13
I will try it with inotify :)