@tiann 2023-02-13
We need this PR : https://github.com/notify-rs/notify/pull/470 to be merged first