@0xpr03 2023-02-13
If you only need inotify, and no other OS support, then notify is probably overkill and inotify-rs could be enough.