@tiann 2023-02-14
And i did a benchmark:

Without inotify:

```
cost: 20.716512ms
cost: 35.034µs
cost: 20.799764ms
cost: 20.795166ms
cost: 20.71757ms
cost: 20.412802ms
cost: 20.288493ms
cost: 20.745402ms
cost: 20.602376ms
cost: 20.689046ms

Average: 18ms
```

With inotify:

```
cost: 11.391805ms
cost: 15.421µs
cost: 6.936767ms
cost: 2.65031ms
cost: 12.248576ms
cost: 8.989787ms
cost: 7.116781ms
cost: 9.271932ms
cost: 15.299µs
cost: 1.726196ms

Average: 6ms
```

Both are release build for arm64-v8a, It is about 12ms faster.