@tiann 2023-02-14
Maybe we change the interval wait time to 1ms without inotify, and it should also work