@tiann 2023-02-20
What should we do now? I changed it to loop wait 1ms