@flxo 2023-02-20
hi @mdaffin. What's your opinion on that?